requirements.txt
Flask
flask-cors
gunicorn
pandas
scikit-learn
numpy
matplotlib
Pillow
opencv-python-headless
keras
tensorflow
fastapi
uvicorn
pybase64
typing
pydantic
torch>=2.0.0
nnunetv2>=2.2
nibabel>=5.0.0
SimpleITK>=2.2.0
scikit-image>=0.21.0